+++ js/data.js
@@ -1,0 +1,55 @@
+const getPhotoData = () => {
+  const DESCRIPTIONS = [
+    'отель',
+    'указатель "go to the beach"',
+    'пляж',
+    'девчуля и фотоаппарат',
+    'весёлый супчик',
+    'суперкар',
+    'клубничка на тарелке',
+    'сочок',
+    'девчуля и самолёт',
+    'обувь',
+    'тропинка на пляже',
+    'ауди',
+    'какая-то еда',
+    'котосуши',
+    'прикольные тапки',
+    'небо над горами',
+    'оркестр',
+    'шеви импала ❤️',
+    'тапки с фонариками',
+    'пальмы',
+    'салатик',
+    'морюшко на закате',
+    'крабик',
+    'концерт 🤘🏻',
+    'страшно, очень страшно',
+  ];
+
+  const NAMES = [
+    'Алексей',
+    'Дарья',
+    'Пётр',
+    'Мария',
+    'Ярослав',
+    'Иван',
+    'Кристина',
+    'Антон',
+    'Андрей',
+    'Татьяна',
+  ];
+
+  const MESSAGES = [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+  ];
+
+  return { DESCRIPTIONS, NAMES, MESSAGES };
+}
+
+export { getPhotoData };
